added comparison mode to matrix, contrib to #67
users may enable comparison mode afterwards it's possible to select
elements from the matrix to be compared. these elements are highlighted
by a background pattern and listed below the matrix. a link is
generated leading to the website that compares selected elements

the current layout probably sucks, but i'm sure gary has some nice ideas
how to enhance it ;-)