implemented js preferences api, see #74
* all preferences are now available through `preferences["key"]`
* preferences can be updated by sending a json object such as the
following to `contextPath + "/myaccount.html"`:

    {
      task: "updatePref",
      prefKey: "someKey",
      prefVal: "someValue"
    }

to see an example have a look at
src/main/webapp/res/js/visualizers/displayPlotFlot/displayPlotFlot.js +230 ff
there i configured the legend-hide-button to also read and write
the preferences. currently this also triggers the notification mech
to display the changes (so the user knows that his prefs are updates).
not sure if that is wanted?