Allow comparing experiments from all versions of a model/protocol (see #44).
This extends the list of experiments returned with a model/protocol to include those with non-latest versions of each protocol/model.
The 'compare' list shown when viewing an entity now has a 'show all versions' toggle button; only the latest are shown by default.
The 'compare all' button compares either just the latest or everything, depending on what is shown.