Allow showing a 'sub matrix' of experiments via the comparison mode (see #67).
You can now do things like selecting a set of models and then working with these easily.
Each sub-matrix has its own URL as well, for convenient bookmarking/sharing.

Note, however, that if you show a sub-matrix with e.g. some models, and click on a protocol from this,
the 'compare all' link when viewing that protocol still shows all models, not those in the sub-matrix.