Work-in-progress looking at responsiveness of batch-submitting new experiments (see issue #45).
Because we wait for a reply from the backend when submitting a new experiment,
the web UI can wait a long time for a suite of experiments to be submitted,
even though it isn't waiting for the check & run tasks actually to execute.
It would be nicer if the communication with the backend could be done in another thread,
and the UI could receive a 'all experiments queued' response quickly.
But there are some errors that are hard to report in this scenario,
because the background thread can't share a database connection.