Store the interface of a protocol in the database (see issue #89).
As part of this, the new interface extraction task (which runs on protocol upload and during
regular cleanUp for protocols not yet analysed) will add an errors.txt file to the protocol
with any errors that were encountered.  Thanks to issue #54 this file is then automatically
displayed when a protocol is viewed.

Still need to make the interface extractor do a full parse of the protocol, so as to pick up
more errors early (rather than only when an experiment is run).  Also still need to use the
interface information in the annotation tool.